Better data plotting, misc hyperparam tweaks, misc improvements.
Throw error when accessing unset fields in ServerAgentSkillBegin packet.
Change how failed skills match up with pending/accepted command buffers.
Trying to go clientless when client connection closes (probably as a result of a crash).
Fix bug regarding normalized cooldown times.
Fix bug regarding visualizing min&max qvalues.
Remove unnecessary convertThenTrain python funcion.
Do not start training until at least a full episode (roughly) is in the replay buffer.
Plot min, mean, and max TD errors.
Support polyak updates at less frequent intervals.
Improve some state machine logic.
Log events and actions in Intelligence acter to debug hang during pvp.

+++ bot/src/rl/observation.hpp
@@ -36,13 +36,20 @@ public:
   uint32_t opponentCurrentMp_;
   uint32_t opponentMaxMp_;
   bool opponentIsKnockedDown_;
-  int hpPotionCount_;
-  std::array<int, 17> remainingTimeOurBuffs_;
-  std::array<int, 17> remainingTimeOpponentBuffs_;
-  // std::array<int, 2> remainingTimeOurDebuffs_;
-  // std::array<int, 2> remainingTimeOpponentDebuffs_;
-  std::array<int, 32> skillCooldowns_;
-  std::array<int, 3> itemCooldowns_;
+  int hpPotionCount_; // TODO: Normalize here rather than in JaxInterface
+
+  // Normalized to [0,1] for buff duration.
+  std::array<float, 17> remainingTimeOurBuffs_;
+  std::array<float, 17> remainingTimeOpponentBuffs_;
+
+  // std::array<float, 2> remainingTimeOurDebuffs_;
+  // std::array<float, 2> remainingTimeOpponentDebuffs_;
+
+  // Normalized to [0,1] for skill cooldown.
+  std::array<float, 32> skillCooldowns_;
+
+  // Normalized to [0,1] for item cooldown.
+  std::array<float, 3> itemCooldowns_;
 };
 
 } // namespace rl
